Moved handleClick to AccountNavigation instead of AccountNavigation file
