+++ resources/assets/components/pages/AccountPage/Account/AccountRoute.js
@@ -11,67 +11,35 @@ import UserPostsQuery from '../Campaigns/UserPostsQuery';
 import DeleteAccountTab from '../Profile/DeleteAccountTab';
 import Subscriptions from '../Subscriptions/Subscriptions';
 import ReferFriendsTab from '../ReferFriends/ReferFriendsTab';
-import {
-  EVENT_CATEGORIES,
-  trackAnalyticsEvent,
-} from '../../../../helpers/analytics';
-
-const handleAccountNavTabClick = tabName => {
-  trackAnalyticsEvent(`clicked_account_navlink_${tabName}}`, {
-    action: 'link_clicked',
-    category: EVENT_CATEGORIES.navigation,
-    label: `account_${tabName}`,
-  });
-};
 
 const AccountRoute = props => (
   <Switch>
     <Route
       path="/us/account/campaigns"
-      onClick={handleAccountNavTabClick('campaigns')}
       render={() => <UserPostsQuery userId={props.userId} />}
     />
     {props.user.hasBadgesFlag ? (
       <Route
         path="/us/account/badges"
-        onClick={handleAccountNavTabClick('badges')}
         render={() => <BadgesTab {...props} />}
       />
     ) : null}
     {featureFlag('volunteer_credits') ? (
-      <Route
-        path="/us/account/credits"
-        onClick={handleAccountNavTabClick('credits')}
-        component={Credits}
-      />
+      <Route path="/us/account/credits" component={Credits} />
     ) : null}
     {featureFlag('cause_preferences') ? (
-      <Route
-        path="/us/account/interests"
-        onClick={handleAccountNavTabClick('interests')}
-        render={() => <Interests />}
-      />
+      <Route path="/us/account/interests" render={() => <Interests />} />
     ) : null}
     <Route
       path="/us/account/subscriptions"
-      onClick={handleAccountNavTabClick('subscriptions')}
       render={() => <Subscriptions {...props} />}
     />
     <Route
       path="/us/account/delete"
-      onClick={handleAccountNavTabClick('delete')}
       render={() => <DeleteAccountTab {...props} />}
     />
-    <Route
-      path="/us/account/refer-friends"
-      onClick={handleAccountNavTabClick('refer-friends')}
-      component={ReferFriendsTab}
-    />
-    <Route
-      path="/us/account"
-      onClick={handleAccountNavTabClick('account')}
-      render={() => <Profile {...props} />}
-    />
+    <Route path="/us/account/refer-friends" component={ReferFriendsTab} />
+    <Route path="/us/account" render={() => <Profile {...props} />} />
   </Switch>
 );
